Fold jsapi.ts into entry.ts and gate checked-in ReScript output on rebuild
The argument-shape adapters (parser, union, optional, assert, ...) existed in
their own module but had exactly one consumer: the entry that re-exported them
under the same names. Declaring them in entry.ts drops the indirection and the
rename-on-export, and takes the bundle total from 24316 to 24255 bytes.

The *.res.mjs files beside each *.res are compiled output, checked in but
regenerated by every job that reads them — so a commit that forgot to
regenerate passed the whole suite while the file in git said something else.
scripts/assert-no-drift.sh fails when a build changes them, wired into the sury
and e2e jobs after their existing build steps.

That check immediately caught real drift: Ppx_General_test.res.mjs still called
Sury.min/Sury.max, names the entry has not exported since the TypeScript
migration, so the file threw on import and its tests never ran. Regenerated
with the ppx built from source; the e2e suite goes from 87 to 103 tests.
