+++ packages/sury/specs/bundleSize.yaml
@@ -1,6 +1,6 @@
 # Minified+gzipped bytes per public export of index.mjs, plus `total` for the whole entry.
 # Generated by `pnpm spec check --write` — every row is measured, so never hand-write one.
-total: 24316
+total: 24255
 exports:
   $Metadata_Id_make: 4135
   $Metadata_get: 4129
@@ -33,46 +33,46 @@ exports:
   boolean: 4416
   brand: 4136
   compactColumns: 12952
-  cuid: 4526
-  date: 4455
+  cuid: 4525
+  date: 4456
   decoder: 4127
   deepStrict: 4239
   deepStrip: 4236
   dict: 12004
-  email: 4571
-  empty: 4827
+  email: 4570
+  empty: 4832
   enableStandardJSONSchema: 5578
   encoder: 4154
-  enum: 12225
+  enum: 12224
   extendJSONSchema: 4211
   float: 4544
-  fromJSONSchema: 18830
+  fromJSONSchema: 18825
   global: 4173
-  gt: 4835
-  gte: 4834
+  gt: 4836
+  gte: 4836
   inputExpression: 4119
   instance: 4297
   int: 4568
   int32: 4568
   is: 4518
   isAsync: 4192
-  isoDateTime: 4562
-  json: 13430
-  jsonString: 13875
-  jsonStringWithSpace: 13891
-  length: 4816
+  isoDateTime: 4560
+  json: 13431
+  jsonString: 13878
+  jsonStringWithSpace: 13894
+  length: 4822
   list: 12520
   literal: 12212
   lt: 4838
-  lte: 4836
-  maxLength: 4814
+  lte: 4837
+  maxLength: 4815
   merge: 12128
   meta: 4265
-  minLength: 4814
+  minLength: 4816
   nan: 4456
   never: 4188
   noValidation: 4136
-  nonEmpty: 4824
+  nonEmpty: 4826
   nullable: 9827
   nullish: 9367
   number: 4544
@@ -98,9 +98,9 @@ exports:
   toJSONSchema: 5550
   trim: 4637
   tuple: 13270
-  uint8Array: 4583
+  uint8Array: 4582
   union: 12218
   unknown: 4120
-  url: 4524
-  uuid: 4542
+  url: 4520
+  uuid: 4541
   void: 4457
